(feat) Create new user through server endpoint
