+++ server/db.js
@@ -1,8 +1,6 @@
 const mongoose = require('mongoose');
 
-mongoose.connect('mongodb://localhost/mindful', {
-  useMongoClient: true
-});
+mongoose.connect('mongodb://localhost/mindful');
 mongoose.Promise = require('bluebird');
 
 
@@ -12,4 +10,20 @@ db.once('open', function() {
   console.log('db connection is open');
 });
 
-module.exports = db;
+const userSchema = mongoose.Schema({
+  username: String,
+  password: String
+});
+
+const sessionSchema = mongoose.Schema({
+  user_id: Number,
+  createdAt: Date,
+  duration: Number,
+  focusTimestamps: Array,
+  maxFocus: Number
+});
+
+module.exports = {
+  User: mongoose.model('User', userSchema),
+  Session: mongoose.model('Session', sessionSchema),
+};
